+++ terraform/test-fixtures/graph-depends-on-count/main.tf
@@ -1,0 +1,6 @@
+resource "aws_instance" "web" {}
+
+resource "aws_instance" "db" {
+	depends_on = ["aws_instance.web"]
+	count = 2
+}